Refactoring pkg structure. Stage 1
