--- cmd/demux/main.go
@@ -1,127 +1,0 @@
-package main
-
-import (
-	"encoding/json"
-	"fmt"
-	"github.com/google/uuid"
-	"github.com/gorilla/mux"
-	"log"
-	"math/rand"
-	"net/http"
-	"runtime"
-	"time"
-)
-
-type job struct {
-	id   uuid.UUID
-	data []int
-}
-
-type jobResponse struct {
-	id  uuid.UUID
-	out int
-}
-
-// swagger:model CommonError
-type CommonError struct {
-	// Status of the error
-	// in: int64
-	Status int64 `json:"status"`
-	// Message of the error
-	// in: string
-	Message string `json:"message"`
-}
-
-// swagger:model JobSubmit
-type jobRequest struct {
-	// Slice of int to sum
-	// in: int64
-	Jobs []int64 `json:"job" validate:"required,gt=1,drive,numeric"`
-}
-
-func main() {
-	jobslist := jobsGenerator()
-	scheduler(jobslist)
-
-	router := mux.NewRouter()
-
-	// Swagger UI from filesystem
-	fs := http.FileServer(http.Dir("./swaggerui"))
-	router.Handle("/swaggerui/{rest}", http.StripPrefix("/swaggerui", fs))
-
-	router.HandleFunc("/", JobSubmit).Methods("POST")
-
-	srv := &http.Server{
-		Handler:      router,
-		Addr:         "127.0.0.1:8080",
-		WriteTimeout: 15 * time.Second,
-		ReadTimeout:  15 * time.Second,
-	}
-
-	log.Fatal(srv.ListenAndServe())
-
-}
-
-// swagger:route POST / admin JobSubmit
-// Submit job
-//
-// security:
-// - apiKey: []
-// responses:
-//
-//	401: CommonError
-func JobSubmit(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	var v jobRequest
-	err := json.NewDecoder(r.Body).Decode(&v)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-	}
-	fmt.Println(v)
-	fmt.Fprintf(w, "Body: %+v", v)
-
-}
-
-func jobsGenerator() []job {
-	var (
-		jb job
-		id uuid.UUID
-	)
-
-	// uuid.EnableRandPool() are not thread-safe and should only be called when there is no possibility that uuid.New()
-	// generation function will be called concurrently.
-	uuid.EnableRandPool()
-
-	jobs := make([]job, 0)
-	for i := 0; i < 100; i++ {
-		id = uuid.New()
-		jb = job{id, []int{rand.Intn(100), rand.Intn(100)}}
-		jobs = append(jobs, jb)
-	}
-	return jobs
-}
-
-func scheduler(joblist []job) {
-
-	ch := make(chan jobResponse)
-	for _, v := range joblist {
-		go worker(ch, v)
-	}
-
-	for i := 0; i < len(joblist); i++ {
-		//There is a jobs response output
-		//fmt.Println(<-ch)
-		<-ch
-	}
-
-}
-
-func worker(ch chan<- jobResponse, jb job) error {
-	resp := jobResponse{id: jb.id, out: 0}
-	for _, v := range jb.data {
-		resp.out += v
-	}
-	runtime.Gosched()
-	ch <- resp
-	return nil
-}
